refactor: :recycle: reorganize CodeEditor and Playground components to use common Header, remove unused Header files
